# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KodeTalent API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type StudentSkills = { [key: string]: number };

export interface Student {
  id: number;
  name: string;
  email: string;
  college: string;
  city: string;
  year: number;
  field: string;
  githubUrl?: string | null;
  overallScore: number;
  xp: number;
  level: number;
  streakCount: number;
  lastActiveDate?: string | null;
  skills: StudentSkills;
  isPro: boolean;
  createdAt: string;
}

export interface CreateStudentBody {
  name: string;
  email: string;
  college: string;
  city: string;
  year: number;
  field: string;
  githubUrl?: string | null;
}

export interface UpdateStudentBody {
  name?: string;
  college?: string;
  city?: string;
  year?: number;
  field?: string;
  githubUrl?: string | null;
  isPro?: boolean;
}

export type DashboardDataQuestProgress = {
  completed: number;
  total: number;
};

export type DashboardDataCollegeRank = {
  rank: number;
  total: number;
  percentile: number;
};

export interface Quest {
  id: number;
  title: string;
  description: string;
  field: string;
  year: number;
  xpReward: number;
  minutes: number;
  whyItMatters?: string;
  howToDoIt?: string;
}

export interface SkillEntry {
  name: string;
  score: number;
}

export interface DashboardData {
  student: Student;
  todayQuest?: Quest | null;
  questProgress: DashboardDataQuestProgress;
  topSkills: SkillEntry[];
  collegeRank: DashboardDataCollegeRank;
  xpToNextLevel: number;
}

export interface CareerWrapped {
  studentId: number;
  name: string;
  college: string;
  streakDays: number;
  xpGained: number;
  scoreStart: number;
  scoreEnd: number;
  indiaPercentile: number;
  month: string;
}

export type StudentQuestStatusStatus =
  (typeof StudentQuestStatusStatus)[keyof typeof StudentQuestStatusStatus];

export const StudentQuestStatusStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface StudentQuestStatus {
  quest: Quest;
  status: StudentQuestStatusStatus;
  completedAt?: string | null;
}

export interface AnalyzeGithubBody {
  githubUrl: string;
  studentId: number;
}

export type SkillsResultSkills = { [key: string]: number };

export interface SkillsResult {
  skills: SkillsResultSkills;
}

export interface GenerateRoadmapBody {
  year: number;
  field: string;
}

export interface RoadmapMilestone {
  title: string;
  description: string;
  year: number;
  order: number;
  whyItMatters?: string;
  howToDoIt?: string;
  estimatedMinutes?: number;
}

export interface RoadmapResult {
  milestones: RoadmapMilestone[];
}

export interface CreateInterviewSessionBody {
  studentId: number;
  company: string;
  round: string;
}

export interface InterviewSession {
  id: number;
  studentId: number;
  company: string;
  round: string;
  questions: string[];
  answers: string[];
  overallScore?: number | null;
  currentQuestion?: string | null;
  questionNumber: number;
  completed: boolean;
  createdAt: string;
}

export interface SubmitAnswerBody {
  answer: string;
}

export interface InterviewQuestionResponse {
  question?: string | null;
  questionNumber: number;
  completed: boolean;
}

export interface QuestionFeedback {
  question: string;
  studentAnswer: string;
  betterAnswer: string;
  score: number;
}

export interface InterviewEvaluation {
  overallScore: number;
  communicationScore: number;
  technicalScore: number;
  confidenceScore: number;
  weakPoint: string;
  strongPoint: string;
  questionFeedback: QuestionFeedback[];
}

export interface CreateTestSessionBody {
  studentId: number;
  testType: string;
  difficulty: string;
}

export interface MCQQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  topic: string;
}

export interface TestSession {
  id: number;
  studentId: number;
  testType: string;
  difficulty: string;
  questions: MCQQuestion[];
  score?: number | null;
  total: number;
  completed: boolean;
  createdAt: string;
}

export interface SubmitTestBody {
  answers: number[];
}

export interface SectionScore {
  topic: string;
  correct: number;
  total: number;
}

export interface TestResult {
  score: number;
  total: number;
  percentage: number;
  sectionBreakdown: SectionScore[];
  weakTopics: string[];
  averageComparison: string;
}

export interface Job {
  id: number;
  companyName: string;
  role: string;
  requiredSkills: string[];
  ctcMin: number;
  ctcMax: number;
  location: string;
  remote: boolean;
}

export interface JobMatch {
  job: Job;
  matchScore: number;
  matchReason: string;
  isLocked: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  studentId: number;
  name: string;
  college: string;
  overallScore: number;
  xp: number;
  level: number;
  streakCount: number;
  isCurrentUser: boolean;
}

export interface CollegeLeaderboardEntry {
  rank: number;
  college: string;
  city: string;
  avgScore: number;
  studentCount: number;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListQuestsParams = {
  field?: string;
  year?: number;
};

export type GetCollegeLeaderboardParams = {
  college: string;
};