# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KodeTalent API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/analyze-github:
    post:
      operationId: analyzeGithub
      tags: [students]
      summary: Analyze GitHub profile and return skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeGithubBody"
      responses:
        "200":
          description: Analyzed skills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsResult"
  /ai/generate-roadmap:
    post:
      operationId: generateRoadmap
      tags: [quests]
      summary: Generate personalized roadmap milestones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRoadmapBody"
      responses:
        "200":
          description: Generated roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapResult"
  /interview/sessions:
    post:
      operationId: createInterviewSession
      tags: [interview]
      summary: Start a new mock interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewSessionBody"
      responses:
        "201":
          description: Created session with first question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interview/sessions/{id}/question:
    post:
      operationId: getNextInterviewQuestion
      tags: [interview]
      summary: Get next interview question based on previous answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: Next question or completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewQuestionResponse"
  /interview/sessions/{id}/evaluate:
    post:
      operationId: evaluateInterview
      tags: [interview]
      summary: Evaluate completed interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview evaluation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewEvaluation"
  /interview/sessions/{id}:
    get:
      operationId: getInterviewSession
      tags: [interview]
      summary: Get interview session details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
...
  /students/{id}/job-matches:
    get:
      operationId: getJobMatches
      tags: [jobs]
      summary: Get job matches for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobMatch"
    post:
      operationId: generateJobMatches
      tags: [jobs]
      summary: Generate/refresh job matches for a student using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated job matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobMatch"
  /leaderboard/college:
    get:
      operationId: getCollegeLeaderboard
      tags: [leaderboard]
      summary: Get college-wise leaderboard
      parameters:
        - name: college
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: College leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  schemas:
...
    AnalyzeGithubBody:
      type: object
      properties:
        githubUrl:
          type: string
        studentId:
          type: integer
      required:
        - githubUrl
        - studentId
    SkillsResult:
      type: object
      properties:
        skills:
          type: object
          additionalProperties:
            type: integer
      required:
        - skills
    GenerateRoadmapBody:
      type: object
      properties:
        year:
          type: integer
        field:
          type: string
      required:
        - year
        - field
    RoadmapMilestone:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        year:
          type: integer
        order:
          type: integer
        whyItMatters:
          type: string
        howToDoIt:
          type: string
        estimatedMinutes:
          type: integer
      required:
        - title
        - description
        - year
        - order
    RoadmapResult:
      type: object
      properties:
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapMilestone"
      required:
        - milestones
    CreateInterviewSessionBody:
      type: object
      properties:
        studentId:
          type: integer
        company:
          type: string
        round:
          type: string
      required:
        - studentId
        - company
        - round
    InterviewSession:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        company:
          type: string
        round:
          type: string
        questions:
          type: array
          items:
            type: string
        answers:
          type: array
          items:
            type: string
        overallScore:
          type: integer
          nullable: true
        currentQuestion:
          type: string
          nullable: true
        questionNumber:
          type: integer
        completed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - studentId
        - company
        - round
        - questions
        - answers
        - questionNumber
        - completed
        - createdAt
    SubmitAnswerBody:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    InterviewQuestionResponse:
      type: object
      properties:
        question:
          type: string
          nullable: true
        questionNumber:
          type: integer
        completed:
          type: boolean
      required:
        - questionNumber
        - completed
    InterviewEvaluation:
      type: object
      properties:
        overallScore:
          type: integer
        communicationScore:
          type: integer
        technicalScore:
          type: integer
        confidenceScore:
          type: integer
        weakPoint:
          type: string
        strongPoint:
          type: string
        questionFeedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
      required:
        - overallScore
        - communicationScore
        - technicalScore
        - confidenceScore
        - weakPoint
        - strongPoint
        - questionFeedback
    QuestionFeedback:
      type: object
      properties:
        question:
          type: string
        studentAnswer:
          type: string
        betterAnswer:
          type: string
        score:
          type: integer
      required:
        - question
        - studentAnswer
        - betterAnswer
        - score
...
    Job:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        role:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        ctcMin:
          type: number
        ctcMax:
          type: number
        location:
          type: string
        remote:
          type: boolean
      required:
        - id
        - companyName
        - role
        - requiredSkills
        - ctcMin
        - ctcMax
        - location
        - remote
    JobMatch:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        matchScore:
          type: integer
        matchReason:
          type: string
        isLocked:
          type: boolean
      required:
        - job
        - matchScore
        - matchReason
        - isLocked
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        studentId:
          type: integer
        name:
          type: string
        college:
          type: string
        overallScore:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streakCount:
          type: integer
        isCurrentUser:
          type: boolean
      required:
        - rank
        - studentId
        - name
        - college
        - overallScore
        - xp
        - level
        - streakCount
        - isCurrentUser
...